// src/books/enums/author.enum.ts
export enum Author {
  // Autores específicos de CMPC e industria forestal
  LUIS_MATTE_LARRAIN = 'Luis Matte Larraín',
  EXPERTOS_INGENIERIA_FORESTAL = 'Expertos en Ingeniería Forestal',
  ECONOMISTAS_AMBIENTALES = 'Economistas Ambientales',
  HISTORIADORES_INDUSTRIALES_CHILENOS = 'Historiadores Industriales Chilenos',
  LIDERES_CMPC = 'Líderes de CMPC',
  INVESTIGADORES_SOSTENIBILIDAD = 'Investigadores en Sostenibilidad',
  SOCIOLOGOS_ORGANIZACIONALES = 'Sociólogos Organizacionales',
  ESPECIALISTAS_INNOVACION_TECNOLOGICA = 'Especialistas en Innovación Tecnológica',
  ACTIVISTAS_AMBIENTALES = 'Activistas Ambientales',
  EDUCADORES_TECNICOS = 'Educadores Técnicos',
}